split Package into Package and PackageMap
